:art: Refactor filename and same conditions
