+++ Sources/Schedules/Alternative.swift
@@ -1,5 +1,5 @@
 //
-//  Alt.swift
+//  Alternative.swift
 //  OperantKit
 //
 //  Created by Yuto Mizutani on 2019/07/23.
