+++ Sources/Schedules/Conjunctive.swift
@@ -1,5 +1,5 @@
 //
-//  Conj.swift
+//  Conjunctive.swift
 //  OperantKit
 //
 //  Created by Yuto Mizutani on 2019/08/25.
