+++ Sources/Schedules/DiscreteTrial.swift
@@ -7,23 +7,19 @@
 
 import RxSwift
 
-public protocol DiscreteTrialCondition {
-    func condition(_ consequence: Consequence, lastTrialValue: Response) -> Bool
-}
-
 public enum TrialType {
     case time(TimeInterval)
     case response(Int)
     case reinforcement(Int)
 
-    var condition: DiscreteTrialCondition {
+    var condition: FinishableCondition {
         switch self {
         case .time(let v):
-            return TrialTimeCondition(v)
+            return FinishByTime(v)
         case .response(let v):
-            return TrialResponseCondition(v)
+            return FinishByResponse(v)
         case .reinforcement(let v):
-            return TrialReinforcementCondition(v)
+            return FinishByReinforcement(v)
         }
     }
 }
@@ -32,11 +28,49 @@ public extension ReinforcementSchedule {
     /// Discrete trial procedure
     ///
     /// - Complexity: O(1)
+    func trials(by condition: FinishableCondition, numberOfTrials: Int) -> DiscreteTrial {
+        return DiscreteTrial(self,
+                             numberOfTrials: numberOfTrials,
+                             currentValue: 0,
+                             condition: condition)
+    }
+
+    /// Discrete trial procedure
+    ///
+    /// - Complexity: O(1)
     func trials(by type: TrialType, numberOfTrials: Int) -> DiscreteTrial {
         return DiscreteTrial(self,
                              numberOfTrials: numberOfTrials,
                              currentValue: 0,
-                             trialType: type)
+                             condition: type.condition)
+    }
+
+    /// Discrete trial procedure with trial time
+    ///
+    /// - Complexity: O(1)
+    func trialTime(_ value: TimeInterval, numberOfTrials: Int) -> DiscreteTrial {
+        return trials(by: .time(value), numberOfTrials: numberOfTrials)
+    }
+
+    /// Discrete trial procedure with trial time
+    ///
+    /// - Complexity: O(1)
+    func trialTime(_ value: Seconds, numberOfTrials: Int) -> DiscreteTrial {
+        return trialTime(.seconds(value), numberOfTrials: numberOfTrials)
+    }
+
+    /// Discrete trial procedure with trial response
+    ///
+    /// - Complexity: O(1)
+    func trialResponse(_ value: Int, numberOfTrials: Int) -> DiscreteTrial {
+        return trials(by: .response(value), numberOfTrials: numberOfTrials)
+    }
+
+    /// Discrete trial procedure with trial response
+    ///
+    /// - Complexity: O(1)
+    func trialReinforcement(_ value: Int, numberOfTrials: Int) -> DiscreteTrial {
+        return trials(by: .reinforcement(value), numberOfTrials: numberOfTrials)
     }
 }
 
@@ -44,9 +78,45 @@ public extension ObservableType where Element == Consequence {
     /// Discrete trial procedure
     ///
     /// - Complexity: O(1)
-    func trials(by type: TrialType, numberOfTrials: Int) -> Observable<Consequence> {
-        return DiscreteTrial(numberOfTrials: numberOfTrials, trialType: type)
+    func trials(by condition: FinishableCondition, numberOfTrials: Int) -> Observable<Consequence> {
+        return DiscreteTrial(numberOfTrials: numberOfTrials, condition: condition)
             .transform(asObservable())
+    }
+
+    /// Discrete trial procedure
+    ///
+    /// - Complexity: O(1)
+    func trials(by type: TrialType, numberOfTrials: Int) -> Observable<Consequence> {
+        return DiscreteTrial(numberOfTrials: numberOfTrials, condition: type.condition)
+            .transform(asObservable())
+    }
+
+    /// Discrete trial procedure with trial time
+    ///
+    /// - Complexity: O(1)
+    func trialTime(_ value: TimeInterval, numberOfTrials: Int) -> Observable<Consequence> {
+        return trials(by: .time(value), numberOfTrials: numberOfTrials)
+    }
+
+    /// Discrete trial procedure with trial time
+    ///
+    /// - Complexity: O(1)
+    func trialTime(_ value: Seconds, numberOfTrials: Int) -> Observable<Consequence> {
+        return trialTime(.seconds(value), numberOfTrials: numberOfTrials)
+    }
+
+    /// Discrete trial procedure with trial response
+    ///
+    /// - Complexity: O(1)
+    func trialResponse(_ value: Int, numberOfTrials: Int) -> Observable<Consequence> {
+        return trials(by: .response(value), numberOfTrials: numberOfTrials)
+    }
+
+    /// Discrete trial procedure with trial reinforcement
+    ///
+    /// - Complexity: O(1)
+    func trialReinforcement(_ value: Int, numberOfTrials: Int) -> Observable<Consequence> {
+        return trials(by: .reinforcement(value), numberOfTrials: numberOfTrials)
     }
 }
 
@@ -62,27 +132,27 @@ public final class DiscreteTrial: DeinitDisposable, TrialStoreableReinforcementS
 
     public let schedule: ScheduleType!
     public let numberOfTrials: Int
-    public let condition: DiscreteTrialCondition
+    public let condition: FinishableCondition
 
     private var currentValue: Int
 
     public init(_ schedule: ScheduleType,
                 numberOfTrials: Int,
                 currentValue: Int = 0,
-                trialType: TrialType) {
+                condition: FinishableCondition) {
         self.schedule = schedule
         self.numberOfTrials = numberOfTrials
         self.currentValue = currentValue
-        self.condition = trialType.condition
+        self.condition = condition
     }
 
     fileprivate init(numberOfTrials: Int,
                      currentValue: Int = 0,
-                     trialType: TrialType) {
+                     condition: FinishableCondition) {
         self.schedule = nil
         self.numberOfTrials = numberOfTrials
         self.currentValue = currentValue
-        self.condition = trialType.condition
+        self.condition = condition
     }
 
     func completeMap(_ source: Observable<Consequence>) -> Observable<Consequence> {
@@ -119,7 +189,7 @@ public final class DiscreteTrial: DeinitDisposable, TrialStoreableReinforcementS
             currentValue += 1
         }
 
-        if condition.condition(consequence, lastTrialValue: lastTrialValue) {
+        if condition.canFinish(consequence, lastValue: lastTrialValue) {
             update(consequence.response)
         }
     }
