+++ Sources/Schedules/Extinction.swift
@@ -1,5 +1,5 @@
 //
-//  EXT.swift
+//  Extinction.swift
 //  OperantKit
 //
 //  Created by Yuto Mizutani on 2018/11/01.
