+++ Sources/Schedules/RandomTime.swift
@@ -1,39 +1,31 @@
 //
-//  RI.swift
+//  RandomTime.swift
 //  OperantKit
 //
-//  Created by Yuto Mizutani on 2018/11/04.
+//  Created by Yuto Mizutani on 2018/11/13.
 //
 
 import RxSwift
 
-@inline(__always)
-func nextRandom(_ value: TimeInterval) -> Milliseconds {
-    return Milliseconds.random(in: 1...value.milliseconds)
-}
-
 public extension ObservableType where Element: ResponseCompatible {
-    /// Random interval schedule
+    /// Random time schedule
     ///
-    /// - important: In order to distinguish from Time schedule, there is a limitation of one or more responses since last time.
     /// - Parameter value: Reinforcement value
     /// - Complexity: O(1)
-    func randomInterval(_ value: TimeInterval) -> Observable<Consequence> {
-        return RI(value).transform(asResponse())
+    func randomTime(_ value: TimeInterval) -> Observable<Consequence> {
+        return RT(value).transform(asResponse())
     }
 }
 
-/// Random interval schedule
+/// Random time schedule
 ///
-/// - important: In order to distinguish from Time schedule, there is a limitation of one or more responses since last time.
 /// - Parameter value: Reinforcement value
-public typealias RI = RandomInterval
+public typealias RT = RandomTime
 
-/// Random interval schedule
+/// Random time schedule
 ///
-/// - important: In order to distinguish from Time schedule, there is a limitation of one or more responses since last time.
 /// - Parameter value: Reinforcement value
-public final class RandomInterval: ResponseStoreableReinforcementSchedule {
+public final class RandomTime: ResponseStoreableReinforcementSchedule {
     public var lastReinforcementValue: Response = .zero
 
     private let value: TimeInterval
@@ -41,7 +33,7 @@ public final class RandomInterval: ResponseStoreableReinforcementSchedule {
 
     public init(_ value: TimeInterval) {
         self.value = value
-        currentRandom = nextRandom(value)
+        self.currentRandom = nextRandom(value)
     }
 
     public convenience init(_ value: Seconds) {
@@ -50,7 +42,7 @@ public final class RandomInterval: ResponseStoreableReinforcementSchedule {
 
     private func outcome(_ response: ResponseCompatible) -> Consequence {
         let current: Response = response.asResponse() - lastReinforcementValue
-        let isReinforcement: Bool = current.numberOfResponses > 0 && current.milliseconds >= currentRandom
+        let isReinforcement: Bool = current.milliseconds >= currentRandom
         if isReinforcement {
             return .reinforcement(response)
         } else {
