+++ Sources/Schedules/VariableInterval.swift
@@ -1,49 +1,72 @@
 //
-//  FI.swift
+//  VariableInterval.swift
 //  OperantKit
 //
-//  Created by Yuto Mizutani on 2018/11/01.
+//  Created by Yuto Mizutani on 2018/11/03.
 //
 
 import RxSwift
 
+@inline(__always)
+func generatedInterval(_ value: TimeInterval, iterations: Int) -> [Milliseconds] {
+    return FleshlerHoffman().generatedInterval(
+        value: value.milliseconds,
+        iterations: iterations
+    )
+}
+
 public extension ObservableType where Element: ResponseCompatible {
-    /// Fixed interval schedule
+    /// Variable interval schedule
     ///
     /// - important: In order to distinguish from Time schedule, there is a limitation of one or more responses since last time.
     /// - Parameter value: Reinforcement value
     /// - Complexity: O(1)
-    func fixedInterval(_ value: TimeInterval) -> Observable<Consequence> {
-        return FI(value).transform(asResponse())
+    func variableInterval(_ value: TimeInterval, iterations: Int = 12) -> Observable<Consequence> {
+        return VI(value, iterations: iterations).transform(asResponse())
+    }
+
+    /// Variable interval schedule
+    ///
+    /// - important: In order to distinguish from Time schedule, there is a limitation of one or more responses since last time.
+    /// - Parameter value: Reinforcement value
+    /// - Complexity: O(1)
+    func variableInterval(_ values: [Milliseconds]) -> Observable<Consequence> {
+        return VI(values).transform(asResponse())
     }
 }
 
-/// Fixed interval schedule
+/// Variable interval schedule
 ///
 /// - important: In order to distinguish from Time schedule, there is a limitation of one or more responses since last time.
 /// - Parameter value: Reinforcement value
-public typealias FI = FixedInterval
+public typealias VI = VariableInterval
 
-/// Fixed interval schedule
+/// Variable interval schedule
 ///
 /// - important: In order to distinguish from Time schedule, there is a limitation of one or more responses since last time.
 /// - Parameter value: Reinforcement value
-public final class FixedInterval: ResponseStoreableReinforcementSchedule {
+public final class VariableInterval: ResponseStoreableReinforcementSchedule {
     public var lastReinforcementValue: Response = .zero
 
-    private let value: TimeInterval
+    private let values: [Milliseconds]
+    private var index: Int = 0
 
-    public init(_ value: TimeInterval) {
-        self.value = value
+    public init(_ values: [Milliseconds]) {
+        self.values = values
     }
 
-    public convenience init(_ value: Seconds) {
-        self.init(TimeInterval.seconds(value))
+    public convenience init(_ value: TimeInterval, iterations: Int = 12) {
+        let values: [Milliseconds] = generatedInterval(value, iterations: iterations)
+        self.init(values)
+    }
+
+    public convenience init(_ value: Seconds, iterations: Int = 12) {
+        self.init(TimeInterval.seconds(value), iterations: iterations)
     }
 
     private func outcome(_ response: ResponseCompatible) -> Consequence {
         let current: Response = response.asResponse() - lastReinforcementValue
-        let isReinforcement: Bool = current.numberOfResponses > 0 && current.milliseconds >= value.milliseconds
+        let isReinforcement: Bool = current.numberOfResponses > 0 && current.milliseconds >= values[index]
         if isReinforcement {
             return .reinforcement(response)
         } else {
@@ -53,6 +76,10 @@ public final class FixedInterval: ResponseStoreableReinforcementSchedule {
 
     public func updateLastReinforcement(_ consequence: Consequence) {
         func update(_ response: ResponseCompatible) {
+            index += 1
+            if index >= values.count {
+                index = 0
+            }
             lastReinforcementValue = response.asResponse()
         }
